move test code into production code since dependencies
