+++ .gitignore
@@ -2,3 +2,4 @@
 .venv/
 src/__pycache__/
 tests/__pycache__/
+.idea
