+++ src/four_in_a_row.py
@@ -1,13 +1,11 @@
 import pygame
 
 from src.messages import LeftMouseDownAt, LeftMouseUpAt, MouseMovedTo, Tick
-from src.states import StartScreenState
-from src.constants import WIDTH, HEIGHT, FPS
+from src.states import StartScreenState, GameOverState, GameState
+from src.constants import WIDTH, HEIGHT, FPS, EMPTY, RED, YELLOW, ROWS, COLUMNS
 
-from src.update import update
+from src.update import update, print_color
 from src.view import view
-
-from tests.test_four_in_a_row import project_model
 
 
 def main():
@@ -126,3 +124,31 @@ if __name__ == '__main__':
 # TODO: ALSA blocks mixer init. Can it be solved without resorting to 'catch exception / null audio api'?
 # TODO: idea to reduce missing updating projectors with new state:
 #         investigate using nametuple string representation in projectors
+def project_model(model):
+    state_string = model.__class__.__name__ + '\n'
+    if isinstance(model, StartScreenState):
+        state_string += f'{model.time=}\n'
+        state_string += f'{model.music_playing=}\n'
+    if isinstance(model, GameOverState):
+        state_string += f'{print_color(model.winner).title()} won.\n'
+    if isinstance(model, GameState):
+        state_string += f'It is {print_color(model.whos_turn_is_it)}s turn.\n'
+        state_string += f'{model.time=}\n'
+        state_string += f'The mouse is at {model.mouse_pos}.\n'
+        state_string += f'{model.mouse_down_time=}\n'
+        state_string += project_board(model.board)
+    return state_string
+
+
+def project_board(board):
+    board_string = ''
+    symbols = {
+        EMPTY: 'O',
+        RED: 'R',
+        YELLOW: 'Y'
+    }
+    board_string += "0 1 2 3 4 5 6\n"
+    board_string += "-------------\n"
+    for y in range(ROWS):
+        board_string += ' '.join(symbols[board[(x, y)]] for x in range(COLUMNS)) + '\n'
+    return board_string
